backend/requirements/backend.txt
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.2.3
Flask-Limiter==3.3.1
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
limits==3.4.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
pip-autoremove==0.10.0
prompt-toolkit==3.0.38
Pygments==2.15.1
pypugjs==5.9.12
pytz==2023.3
pyyaml==6.0
redis==4.5.5
requests==2.30.0
rich==13.3.5
six==1.16.0
typing_extensions==4.5.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
